refactor(delete user): Changed directory location
