+++ src/components/user/deleteUser/DeleteUser.jsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { Formik, Form, Field } from 'formik';
 import * as Yup from 'yup';
 import { Navigate, useNavigate } from 'react-router-dom';
-import useAuth from '../../hooks/useAuth';
+import useAuth from '../../../hooks/useAuth';
 
 function DeleteUser({ userId }) {
   const [loading, setLoading] = useState(false);
