+++ src/components/userForm/UserForm.jsx
@@ -10,7 +10,7 @@ import * as Yup from 'yup';
 import useAuth from '../../hooks/useAuth';
 import './UserForm.css';
 import ButtonBack from '../buttons/buttonBack/ButtonBack';
-import DeleteUser from '../deleteUser/DeleteUser';
+import DeleteUser from '../user/deleteUser/DeleteUser';
 
 function UserForm() {
   const [loading, setLoading] = useState(false);
@@ -210,7 +210,6 @@ function UserForm() {
             const newUser = respuesta.user;
             const successMessage = isUpdating ? 'Usuario modificado satisfactoriamente' : 'Usuario creado satisfactoriamente';
 
-            // TMBN FALTA ESO DE QUE EL EDITAR UN NO ADMIN SE MUERE
             if ((id == currentUser?.id) || !id) {
               newUser.token = respuesta.token ? respuesta.token : currentUser.token;
               handleUserLogin(newUser);
